+++ modules/ingress/rest_subscriber/src/services/EmpathicBuildingService.ts
