+++ modules/ingress/rest_subscriber/src/utils/eb_types.ts
