+++ packages/ingress/src/index.ts
@@ -1,14 +1,5 @@
-import { EmpathicBuildingService } from "./EmpathicBuilding";
 import { RESTService } from "./REST";
 
-import type { EmpathicBuildingConfig, DecodedEvent } from "./EmpathicBuilding/types";
-
 export {
-    EmpathicBuildingService,
     RESTService,
-};
-
-export type {
-    EmpathicBuildingConfig,
-    DecodedEvent,
 };
